refactor(m9): extraire HeroDetailContent — composant dédié #266
- Crée src/components/HeroDetailContent.tsx
- HeroUpgradeModal.tsx utilise le nouveau composant
- Zéro changement de comportement visible
